refactor(viewmodel): extract InternetRadioViewModel models
Lone data class lifted to InternetRadioViewModelModels.kt. The
InternetRadioViewModel class itself remains a single oversized unit
(~900 LOC) and needs collaborator extraction (RadioStationRepository,
RadioPlaybackController, RadioFavoritesManager) in a follow-up PR.
